+++ lib/src/interop/style/sources/geojson_source_interop.dart
@@ -12,4 +12,7 @@ class GeoJsonSourceJsImpl {
     String type,
     FeatureCollectionJsImpl data,
   });
+
+  external GeoJsonSourceJsImpl setData(
+      FeatureCollectionJsImpl featureCollection);
 }